+++ Gemfile
@@ -4,3 +4,4 @@ gem 'pry'
 gem 'ruby-progressbar'
 gem 'rainbow'
 gem 'thor'
+gem 'minitest'
