+++ README.md
@@ -7,6 +7,8 @@
 
 2015-4-25
 
+> 添加了单元测试
+
 > 游戏运行方式修改为标准CLI交互，于是请这样进入游戏。
 
 ```
